// src/app/core/models/resumen.simulacion.model.ts
export interface ResumenItem {
  votanteId: number;
  jnE_Id: number;
  organizacionId: number;
  organizacionPolitica: string;
  logoUrl: string;
  eleccionId: number;
  fechaHora: string;
  listaId: number;
  preferencia1: number;
  preferencia2: number;
  nombres: string;
  apellidos: string;
  fotoUrl: string;
  tipoCamara: string;
}